Refactor Visualization
Move _render_plot, etc. to separate internal module _plots.
